# Make tokenQueue a plain slice