As per title. This makes the code a bit more simple and avoids a few allocations. YAY?

## Benchmark

```
benchmark                                                                                       old ns/op     new ns/op     delta
BenchmarkUnmarshal/struct_data_v0.3-16                                                          17996         17630         -2.03%
BenchmarkUnmarshal/string_data_v0.3-16                                                          17431         17137         -1.69%
BenchmarkUnmarshal/data,_attributes_and_extensions_and_specversion_with_struct_data_v0.3-16     19363         18604         -3.92%
BenchmarkUnmarshal/base64_json_encoded_data_v1.0-16                                             12238         11992         -2.01%
BenchmarkUnmarshal/nil_data_v1.0-16                                                             16317         16048         -1.65%
BenchmarkUnmarshal/nil_data_v0.3-16                                                             16464         16268         -1.19%
BenchmarkUnmarshal/struct_data_v1.0-16                                                          17902         17714         -1.05%
BenchmarkUnmarshal/data,_attributes_and_extensions_and_specversion_with_struct_data_v1.0-16     19159         18351         -4.22%
BenchmarkUnmarshal/string_data_v1.0-16                                                          17243         17140         -0.60%
BenchmarkUnmarshal/base64_xml_encoded_data_v1.0-16                                              12775         12725         -0.39%

benchmark                                                                                       old allocs     new allocs     delta
BenchmarkUnmarshal/struct_data_v0.3-16                                                          79             78             -1.27%
BenchmarkUnmarshal/string_data_v0.3-16                                                          78             77             -1.28%
BenchmarkUnmarshal/data,_attributes_and_extensions_and_specversion_with_struct_data_v0.3-16     88             87             -1.14%
BenchmarkUnmarshal/base64_json_encoded_data_v1.0-16                                             52             51             -1.92%
BenchmarkUnmarshal/nil_data_v1.0-16                                                             74             73             -1.35%
BenchmarkUnmarshal/nil_data_v0.3-16                                                             74             73             -1.35%
BenchmarkUnmarshal/struct_data_v1.0-16                                                          77             76             -1.30%
BenchmarkUnmarshal/data,_attributes_and_extensions_and_specversion_with_struct_data_v1.0-16     88             87             -1.14%
BenchmarkUnmarshal/string_data_v1.0-16                                                          76             75             -1.32%
BenchmarkUnmarshal/base64_xml_encoded_data_v1.0-16                                              52             51             -1.92%

benchmark                                                                                       old bytes     new bytes     delta
BenchmarkUnmarshal/struct_data_v0.3-16                                                          5263          4750          -9.75%
BenchmarkUnmarshal/string_data_v0.3-16                                                          5247          4734          -9.78%
BenchmarkUnmarshal/data,_attributes_and_extensions_and_specversion_with_struct_data_v0.3-16     6537          6024          -7.85%
BenchmarkUnmarshal/base64_json_encoded_data_v1.0-16                                             4387          3874          -11.69%
BenchmarkUnmarshal/nil_data_v1.0-16                                                             5031          4518          -10.20%
BenchmarkUnmarshal/nil_data_v0.3-16                                                             5031          4518          -10.20%
BenchmarkUnmarshal/struct_data_v1.0-16                                                          5087          4574          -10.08%
BenchmarkUnmarshal/data,_attributes_and_extensions_and_specversion_with_struct_data_v1.0-16     6537          6024          -7.85%
BenchmarkUnmarshal/string_data_v1.0-16                                                          5071          4558          -10.12%
BenchmarkUnmarshal/base64_xml_encoded_data_v1.0-16                                              4403          3890          -11.65%

```